move pinned buffers
