+++ lua/vuffers/buffers/init.lua
@@ -1,4 +1,5 @@
 local bufs = require("vuffers.buffers.buffers")
-local M = vim.tbl_deep_extend("force", {}, bufs)
+local pinned = require("vuffers.buffers.pinned-buffers")
+local M = vim.tbl_deep_extend("force", {}, bufs, pinned)
 
 return M
