+++ tests/buffers_spec.lua
@@ -23,7 +23,7 @@ describe("buffers >>", function()
   describe("pin_buffer >>", function()
     before_each(function()
       event_bus._delete_all_subscriptions()
-      buffers._reset_buffers()
+      buffers.set_buffers({})
     end)
 
     it("should pin buffer", function()
@@ -79,7 +79,7 @@ describe("buffers >>", function()
   describe("unpin_buffer >>", function()
     before_each(function()
       event_bus._delete_all_subscriptions()
-      buffers._reset_buffers()
+      buffers.set_buffers({})
     end)
 
     it("should unpin buffer", function()
@@ -140,7 +140,7 @@ describe("buffers >>", function()
   describe("remove_unpinned_buffers >>", function()
     before_each(function()
       event_bus._delete_all_subscriptions()
-      buffers._reset_buffers()
+      buffers.set_buffers({})
     end)
 
     it("should remove unpinned buffers when active buffer is pinned", function()
